fix target loadinng bug

+++ validation.py
@@ -36,7 +36,7 @@ def val_epoch(epoch, data_loader, model, criterion, opt, logger, writer):
             data_time.update(time.time() - end_time)
 
             inputs = data[0]
-            targets = data[1].float()
+            targets = data[1]
 
             # all_targets.append(targets)
